feat: healthkit 마무리 작업
