+++ Projects/App/WatchExtension/Sources/App/RootApp.swift
@@ -10,31 +10,37 @@ import ComposableArchitecture
 
 @main
 struct RootApp: App {
-    var body: some Scene {
+    @SceneBuilder var body: some Scene {
         WindowGroup {
             RootView(store: .init(initialState: RootStore.State(), reducer: RootStore()._printChanges()))
+                .onAppear {
+                    HealthKitManager.shared.requestAuthorization() { _, _ in
+
+                    }
+                }
         }
     }
 }
 
 //TODO: workout session 으로 마이그레이션 후 주석 제거
-/*
- struct WatchApp: App {
- @StateObject private var workoutManager = WorkoutManager()
- 
- @SceneBuilder var body: some Scene {
- WindowGroup {
- NavigationStack {
- StartView()
- }
- .onAppear {
- workoutManager.requestAuthorization()
- }
- .sheet(isPresented: $workoutManager.showingSummaryView) {
- SummaryView()
- }
- .environmentObject(workoutManager)
- }
- }
- }
- */
+
+//@main
+//struct RootApp: App {
+//    @StateObject private var workoutManager = WorkoutManager()
+//
+//    @SceneBuilder var body: some Scene {
+//        WindowGroup {
+//            NavigationStack {
+//                StartView()
+//            }
+//            .onAppear {
+//                workoutManager.requestAuthorization()
+//            }
+//            .sheet(isPresented: $workoutManager.showingSummaryView) {
+//                SummaryView()
+//            }
+//            .environmentObject(workoutManager)
+//        }
+//    }
+//}
+//
