+++ Projects/App/WatchExtension/Sources/Domain/WorkoutClient.swift
@@ -1,0 +1,8 @@
+//
+//  WorkoutClient.swift
+//  AppWatchExtension
+//
+//  Created by 송영모 on 2023/06/29.
+//
+
+import Foundation
