+++ Projects/App/WatchExtension/Sources/Feature/HomeStore.swift
@@ -16,6 +16,10 @@ public struct HomeStore: ReducerProtocol {
     
     public struct State: Equatable {
         public weak var watchConnectivityDelegate: HomeWatchConnectivityDelegate?
+        public weak var workoutDelegate: HomeWorkoutDelegate?
+        
+        public var heartRate: Int = 0
+        public var calorie: Int = 0
         
         public init() { }
     }
@@ -24,9 +28,14 @@ public struct HomeStore: ReducerProtocol {
         case binding(BindingAction<State>)
         
         case setWatchConnectivityDelegate(HomeWatchConnectivityDelegate)
+        case setWorkoutDelegate(HomeWorkoutDelegate)
         
-        case test1ButtonTapped
-        case test2ButtonTapped
+        case startButtonTapped
+        
+        case setHeartRate(Int)
+        case setCalorie(Int)
+
+        case sendMessage(String, Int)
     }
     
     public var body: some ReducerProtocol<State, Action> {
@@ -39,12 +48,13 @@ public struct HomeStore: ReducerProtocol {
                 
             case let .setWatchConnectivityDelegate(watchConnectivityDelegate):
                 state.watchConnectivityDelegate = watchConnectivityDelegate
+                //TODO: 로직 수정 완료 후 제거
+                /*
                 HealthKitManager.shared.requestObserverQuery() { [watchConnectivityDelegate] (query, samplesOrNil, deletedObjectsOrNil, newAnchor, errorOrNil) in
                     guard let samples = samplesOrNil, let sample = samples.last as? HKQuantitySample else {
                         return
                     }
                     
-                    print(sample)
                     
                     if sample.sampleType == .quantityType(forIdentifier: .heartRate) {
                         watchConnectivityDelegate.session?.sendMessage(
@@ -54,14 +64,27 @@ public struct HomeStore: ReducerProtocol {
                         watchConnectivityDelegate.session?.sendMessage(["message" : sample.quantity.doubleValue(for: HKUnit.kilocalorie())], replyHandler: nil)
                     }
                 }
-                
+                 */
                 return .none
                 
-            case .test1ButtonTapped:
-                debugPrint("1")
+            case let .setWorkoutDelegate(workoutDelegate):
+                state.workoutDelegate = workoutDelegate
                 return .none
                 
-            default:
+            case .startButtonTapped:
+                state.workoutDelegate?.startWorkout(workoutType: .functionalStrengthTraining)
+                return .none
+                
+            case let .setHeartRate(heartRate):
+                state.heartRate = heartRate
+                return .send(.sendMessage("heartRate", heartRate))
+                
+            case let .setCalorie(calorie):
+                state.calorie = calorie
+                return .send(.sendMessage("calorie", calorie))
+
+            case let .sendMessage(key, value):
+                state.watchConnectivityDelegate?.session?.sendMessage([key: value], replyHandler: nil)
                 return .none
             }
         }
